requirements.txt
PyQt5==5.15.10
paramiko==3.4.0
PyYAML==6.0.2
APScheduler==3.10.4
cryptography==42.0.7
